# High performance `getledgerentry`