--- a/docs/stellar-core_example.cfg
+++ b/docs/stellar-core_example.cfg
@@ -58,6 +58,7 @@ PREFETCH_BATCH_SIZE=1000
 # HTTP_PORT (integer) default 11626
 # What port stellar-core listens for commands on.
 # If set to 0, disable HTTP interface entirely
+# Must not be the same as HTTP_QUERY_PORT if not 0.
 HTTP_PORT=11626
 
 # PUBLIC_HTTP_PORT (true or false) default false
@@ -75,6 +76,27 @@ HTTP_MAX_CLIENT=128
 COMMANDS=[
 "ll?level=info&partition=Herder"
 ]
+
+# HTTP_QUERY_PORT (integer) default 0
+# What port stellar-core listens for query commands on,
+# such as getledgerentryraw.
+# If set to 0, disable HTTP query interface entirely.
+# Must not be the same as HTTP_PORT if not 0.
+HTTP_QUERY_PORT=0
+
+# QUERY_THREAD_POOL_SIZE (integer) default 4
+# Number of threads available for processing query commands.
+# If HTTP_QUERY_PORT == 0, this option is ignored.
+QUERY_THREAD_POOL_SIZE=4
+
+# QUERY_SNAPSHOT_LEDGERS (integer) default 0
+# Number of historical ledger snapshots to maintain for
+# query commands. Note: Setting this to large values may
+# significantly impact performance. Additionally, these
+# snapshots are a "best effort" only and not persisted on
+# restart. On restart, only the current ledger will be
+# available, with snapshots avaiable as ledgers close.
+QUERY_SNAPSHOT_LEDGERS = 0
 
 # convenience mapping of common names to node IDs. The common names can be used
 #  in the .cfg. `$common_name`. If set, they will also appear in your logs