--- a/lib/httpthreaded/request.hpp
+++ b/lib/httpthreaded/request.hpp
@@ -28,6 +28,7 @@ struct request
     int http_version_major;
     int http_version_minor;
     std::vector<header> headers;
+    std::string body{};
 };
 
 } // namespace server