--- a/make-mks
+++ b/make-mks
@@ -59,7 +59,7 @@ listall() {
  echo SOCI_PG_FILES = $(listall soci/src/backends/postgresql)
 
  echo SQLITE3_FILES = $(listall sqlite)
- echo UTIL_FILES = $(listall util http)
+ echo UTIL_FILES = $(listall util http httpthreaded)
  echo ASIO_H_FILES = $(listall asio/asio/include)
  # The following does not work without boost or -DASIO_STANDALONE=1
  #echo ASIO_CXX_FILES = asio/src/*.cpp