# Description

Resolves #4306

Note: the following interface is now outdated. Please refer to `docs/software/commands.md` for the up to date interface. The performance measurements are still accurate.

Previous interface:

`getledgerentry` core endpoint is now high performance, non blocking and served by a multi-threaded HTTP server that does not interact with the main thread. This enables down stream systems to query this endpoint at high rates without `captive-core` nodes losing sync. Note that this endpoint is served by a different port and separated from stellar-core's other endpoints. The following config options have been added supporting this feature:

```
RPC_HTTP_PORT = 11627 # default listening port
RPC_THREADS = 4 # default threads serving getledgerentry endpoint
RPC_SNAPSHOT_LEDGERS = 5 # default number of ledgers retained in history
```
The HTTP request string is as follows:

```
getledgerentry?key=Base64&ledgerSeq=NUM
```

`key` is required, and is the Base64 XDR of the LedgerKey being queried. `ledgerSeq` is optional. If not set, stellar-core will return the LedgerEntry based on the most recent ledger. If `ledgerSeq` is set, stellar-core will return an entry based on a historical ledger snapshot at the given ledger. The return payload is a JSON object in the following format:

```
"ledger": ledgerSeq, // Required
"state": ["not_found" | "live" | "dead"], // Required
"entry": Base64 // Optional
```

`ledger` is the ledgerSeq that the query is based on, and is always returned. `state` returns "live" if a live LedgerEntry was found, or "dead" if the LedgerEntry does not exist. Additionally, if `ledgerSeq` is set to a snapshot that stellar-core does not currently have, "not_found" is returned. Finally, if `state==live`, "entry" is returned with the Base64 XDR encoding of the full LedgerEntry.

To measure performance, I used a parallel go script  (thanks @Shaptic) with `stellar/go/clients/stellarcore` to send requests at a very high rate over local host. 1 million LedgerEntries of type ContractCode, ContractData, and Trustline were randomly sampled from the BucketList (such that all entries exist) for these requests, and no caching was used. Test was ran on `test-core-003a.dev.stellar002` with a `captive-core` instance in sync with pubnet with the following benchmarks:

```
Request Rate: 2731 / sec
total queries: 100910, failed: 0, success: 100910
success rate: 100.000000
average latency: 366.163µs
min latency: 256.861µs

Request Rate: 11836 / sec
total queries: 103390, failed: 0, success: 103390
success rate: 100.000000
average latency: 422.453µs
min latency: 204.412µs

Request Rate: 18245 / sec
total queries: 100740, failed: 0, success: 100740
success rate: 100.000000
average latency: 548.105µs
min latency: 214.905µs
```

# Checklist
- [x] Reviewed the [contributing](https://github.com/stellar/stellar-core/blob/master/CONTRIBUTING.md#submitting-changes) document
- [x] Rebased on top of master (no merge commits)
- [x] Ran `clang-format` v8.0.0 (via `make format` or the Visual Studio extension)
- [x] Compiles
- [x] Ran all tests
- [ ] If change impacts performance, include supporting evidence per the [performance document](https://github.com/stellar/stellar-core/blob/master/performance-eval/performance-eval.md)