@SirTyson 2024-06-19
I've now added a batch load endpoint called `getledgerentrybatch`. This is a POST method and requires the following body:
```
ledgerSeq=NUM&key=Base64&key=Base64...
```

`ledgerSeq` is an optional value that follows the same semantics as the `getledgerentry` endpoint. It is followed by one or more `key` to be queried. The return value is a JSON payload as follows:
```
{
"entries": [
  {"entry": "Base64-LedgerKey", "state": "dead"}, // dead entry
  {"entry": "Base64-LedgerEntry", "state": "live"}, // live entry
],
"ledger": ledgerSeq
}
```

If a ledgerSeq is queried but is not available, the return payload is as follows:
```
{"ledger": ledgerSeq, "state": "not_found"}
```