@2opremio 2024-06-19
I know this is just a prototype, but it would be more ergonomic to send JSON.

Also, it seems like entry could be both an entry or a key, I would suggest always providing a key and optionally providing an entry field which can be omitted:

```json
{
"entries": [
  {"key": "Base64-LedgerKey", "state": "dead"}, // dead entry
  {"key": "Base64-LedgerKey", "entry": "Base64-LedgerEntry", "state": "live"}, // live entry
],
"ledger": ledgerSeq
}
```