@2opremio 2024-06-19
Additionally, how can I distinguish TTL'ed entries? Are TTL'ed entries the ones with `"dead"` status?

I will need the entry for those as well.