@2opremio 2024-06-19
Ah, ok, so live means not evicted, and I need to query TTL entries separately (the more reason to have a batch endpoint)