@2opremio 2024-06-19
If `"dead"` means not found, we can simply omit those entries and assume omitted entries where not found.

Then we can get rid of the state field altogether.