@SirTyson 2024-08-07
> What is the maximum number of ledger entry keys this endpoint could accept?

I haven't tested the maximum entries for a single query. However I doubt it will be a limiting factor, given we achieved a request rate of 20k RPS with an average latency of 548.105 us for point loads, and bulk loads are more efficient.