@janewang 2024-08-07
@SirTyson We currently support 200 keys which is more than sufficient. 
Btw the current interface is here and I hope this is a non-breaking change for downstream clients: https://developers.stellar.org/docs/data/rpc/api-reference/methods/getLedgerEntries